+++ paper/paper.bib
@@ -4,7 +4,7 @@
     title = {Benefits and limitations of genome- wide association studies},
   journal = {Nat Rev Genet},
      year = 2019,
-   volume = 20(8),
+   volume = "20(8)",
     pages = {467-484},
       doi = {10.1038/s41576-019-0127-1},
    }
