+++ paper/paper.bib
@@ -14,7 +14,7 @@
     title = {Odyssey: a semi-automated pipeline for phasing, imputation, and analysis of genome-wide genetic data},
   journal = {BMC Bioinformatics},
      year = 2019,
-   volume = 20(1),
+   volume = "20(1)",
     pages = {364},
       doi = {10.1186/s12859-019-2964-5},
    }
@@ -24,7 +24,7 @@
     title = {Nextflow enables reproducible computational workflows},
   journal = {Nat. Biotechnol},
      year = 2017,
-   volume = 35(4),
+   volume = "35(4)",
     pages = {316-319},
       doi = {10.1038/nbt.3820},
    }
@@ -34,7 +34,7 @@
     title = {SeqArray-a storage-efficient high-performance data format for WGS variant calls},
   journal = {Bioinformatics},
      year = 2017,
-   volume = 33(15),
+   volume = "33(15)",
     pages = {2251-2257},
       doi = {10.1093/bioinformatics/btx145},
    }
@@ -44,7 +44,7 @@
     title = {Robust inference of population structure for ancestry prediction and correction of stratification in the presence of relatedness},
   journal = {Genet Epidemiol},
      year = 2015,
-   volume = 39(4),
+   volume = "39(4)",
     pages = {276-293},
       doi = {10.1002/gepi.21896},
    }
@@ -54,7 +54,7 @@
     title = {Model-free Estimation of Recent Genetic Relatedness},
   journal = {Am J Hum Genet},
      year = 2016,
-   volume = 98(1),
+   volume = "98(1)",
     pages = {127-148},
       doi = {10.1016/j.ajhg.2015.11.022},
    }   
@@ -64,7 +64,7 @@
     title = {Genetic association testing using the GENESIS R/Bioconductor package},
   journal = {Bioinformatics},
      year = 2019,
-   volume = 35(24),
+   volume = "35(24)",
     pages = {5346-5348},
       doi = {10.1093/bioinformatics/btz567},
    }     
@@ -74,7 +74,7 @@
     title = {Control for Population Structure and Relatedness for Binary Traits in Genetic Association Studies via Logistic Mixed Models},
   journal = {Am J Hum Genet},
      year = 2016,
-   volume = 98(4),
+   volume = "98(4)",
     pages = {653-666},
       doi = {10.1016/j.ajhg.2016.02.012},
    }
