+++ paper/paper.bib
@@ -1,80 +1,88 @@
 @article{Tam:2019,
 	author = {{Tam V} and {Patel N} and {Turcotte M} and
 	{Bossé Y} and {Paré G} and {Meyre D}},
-    title = {Benefits and limitations of genome- wide association studies},
+    title = "{Benefits and limitations of genome- wide association studies}",
   journal = {Nat Rev Genet},
      year = 2019,
-   volume = "20(8)",
+   volume = {20},
+   number = {8},
     pages = {467-484},
       doi = {10.1038/s41576-019-0127-1},
    }
 
 @article{Eller:2019,
 	author = {{Eller RJ} and {Janga SC} and {Walsh S}},
-    title = {Odyssey: a semi-automated pipeline for phasing, imputation, and analysis of genome-wide genetic data},
+    title = "{Odyssey: a semi-automated pipeline for phasing, imputation, and analysis of genome-wide genetic data}",
   journal = {BMC Bioinformatics},
      year = 2019,
-   volume = "20(1)",
+   volume = {20},
+   number = {1},
     pages = {364},
       doi = {10.1186/s12859-019-2964-5},
    }
 
 @article{Di:2017,
 	author = {{Di Tommaso P} and {Chatzou M} and {Floden EW} and {Barja PP} and {Palumbo E} and {Notredame C}},
-    title = {Nextflow enables reproducible computational workflows},
+    title = "{Nextflow enables reproducible computational workflows}",
   journal = {Nat. Biotechnol},
      year = 2017,
-   volume = "35(4)",
+   volume = {35},
+   number = {4},
     pages = {316-319},
       doi = {10.1038/nbt.3820},
    }
 
 @article{Zheng:2017,
 	author = {{Zheng X} and {Gogarten SM} and {Lawrence M} and {et al}},
-    title = {SeqArray-a storage-efficient high-performance data format for WGS variant calls},
+    title = "{SeqArray-a storage-efficient high-performance data format for WGS variant calls}",
   journal = {Bioinformatics},
      year = 2017,
-   volume = "33(15)",
+   volume = {33},
+   number = {15},
     pages = {2251-2257},
       doi = {10.1093/bioinformatics/btx145},
    }
 
 @article{Conomos:2015,
 	author = {{Conomos MP} and {Miller MB} and {Thornton TA}},
-    title = {Robust inference of population structure for ancestry prediction and correction of stratification in the presence of relatedness},
+    title = "{Robust inference of population structure for ancestry prediction and correction of stratification in the presence of relatedness}",
   journal = {Genet Epidemiol},
      year = 2015,
-   volume = "39(4)",
+   volume = {39},
+   number = {4},
     pages = {276-293},
       doi = {10.1002/gepi.21896},
    }
    
 @article{Conomos:2016,
 	author = {{Conomos MP} and {Reiner AP} and {Weir BS} and {Thornton TA}},
-    title = {Model-free Estimation of Recent Genetic Relatedness},
+    title = "{Model-free Estimation of Recent Genetic Relatedness}",
   journal = {Am J Hum Genet},
      year = 2016,
-   volume = "98(1)",
+   volume = {98},
+   number = {1},
     pages = {127-148},
       doi = {10.1016/j.ajhg.2015.11.022},
    }   
    
 @article{Gogarten:2019,
 	author = {{Gogarten SM} and {Sofer T} and {Chen H} and {et al}},
-    title = {Genetic association testing using the GENESIS R/Bioconductor package},
+    title = "{Genetic association testing using the GENESIS R/Bioconductor package}",
   journal = {Bioinformatics},
      year = 2019,
-   volume = "35(24)",
+   volume = {35},
+   number = {24},
     pages = {5346-5348},
[…]
