### 2020-12-01 #2857 @lpantano
Just a comment, the stats of whedon is missing the *nf files that are the main code in this kind of pipelines (nextflow).
