### 2020-12-21 #2857 @ZeyuanSong
Yes, we would like to suggest possible reviewers:
Dr. Riccardo Bellazzi – University of Pavia, Italy – email: riccardo.bellazzi@unipv.it
Dr. Alberto Riva – University of Florida – ariva@ufl.edu,
Dr. Aldi Kraja – University of Mississippi – akraja@umc.edu,
Dr. Han Chen – The University of Texas Health Science Center at Houston – Han.Chen.2@uth.tmc.edu
@lpantano Thank you!
