### 2021-01-14 #2957 @whedon
[REVIEW]: nf-gwas-pipeline: A Nextflow Genome-Wide Association Study Pipeline
**Submitting author:** @ZeyuanSong (<a href="http://orcid.org/0000-0002-7352-4177">Zeyuan Song</a>)
**Repository:** <a href="https://github.com/montilab/nf-gwas-pipeline" target ="_blank">https://github.com/montilab/nf-gwas-pipeline</a>
**Version:** v1.0.0

@preetida,@rspirgel please carry out your review in this issue by updating the checklist below. If you cannot edit the checklist please:
