### 2021-02-02 montilab/nf-gwas-pipeline#2 @rspirgel
Documentation: Community guidelines missing
Did not see mention in the documentation or paper of how users could: 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

https://github.com/openjournals/joss-reviews/issues/2957
