### 2021-02-02 #2857 @preetida
Hello,
Thank you for submitting your software. I had a chance to test the pipeline.
Docker set up works fine, and example works as demonstrated.
However, for running the pipeline on an HPC that does not support docker, following the instruction, running tools gives error for missing dependency.
For e.g. below
```
module load R/3.6.1
module load vcftools
 module load bcftools
module load annovar
module load pandoc
module load plink
nextflow gwas.nf -c gwas.config
Error executing process > 'vcf_to_gds (chr_2)'

Caused by:
  Process `vcf_to_gds (chr_2)` terminated with an error exit status (1)

Command executed:

  02_vcf_to_gds.R chr_2_qc2.vcf.gz chr_2.gds chr_2_vcf_to_gds.log

Command exit status:
  1

Command output:
  [1] "Tue Feb  2 12:09:36 2021"

Command error:
  Error in library(SeqArray) : there is no package called ‘SeqArray’
  Calls: suppressPackageStartupMessages -> withCallingHandlers -> library
  Execution halted

Work dir:
  /home/pjb40/homework/Nextflow/nf-gwas-pipeline/work/d2/c0e24eb2fdda9079d9772828ee8444

Tip: view the complete command output by changing to the process work dir and entering the command `cat .command.out`
WARN: Killing pending tasks (7)
```
