### 2021-02-02 #2857 @anfederico
Hello, co-author here. For HPC's without Docker, an alternative route is to use Singularity. One can convert from the Docker image directly (assuming Singularity is available and setup to pull from Docker).

```
singularity pull docker://montilab/gwas
```

```
nextflow gwas.nf -c gwas.config -with-singularity [singularity image file]
```
