### 2021-02-02 #2857 @lpantano
@anfederico, thanks for helping with this. @preetida try to see if this works. Meanwhile, @anfederico it would be good to have a complete list of dependencies for cases where tools needs to be installed (maybe a conda env file helps to deal with this? or list all the dependencies.), thanks!.
