### 2021-02-02 a5cac18
Update README.md
add R packages required
