### 2021-02-04 c76af3b
change combine.dat to out.dat to match dimensions
