### 2021-02-06 1e1dc90
modified to take flexible chromosome numbers
