### 2021-02-06 23dac38
modified to take flexible chromosome numbers
