### 2021-02-07 7768a55
make the p-value threshold numeric
