### 2021-02-07 80adc9c
Update 06_annovar_input.R
