### 2021-02-07 montilab/nf-gwas-pipeline#2 @anfederico
I've added a contributing document to the repository which describes guidelines for users to adhere to as well as channels (Github Issues) to report issues/seek help. We will refer to this document in the paper.
