### 2021-02-25 #2957 @preetida
@lpantano Yes, the guidelines have been added to the paper. Looks all set.
