### 2021-02-26 #2957 @anfederico
> * [x]  Make a tagged release of your software, and list the version tag of the archived version here.

v.1.00 (Publication Release) tagged on Github
https://github.com/montilab/nf-gwas-pipeline/releases/tag/v1.00

> * [x]  Archive the reviewed software in Zenodo or a similar service (e.g., figshare, an institutional repository)
> * [x]  Check the archival deposit (e.g., in Zenodo) has the correct metadata. This includes the title (should match the paper title) and author list (make sure the list is correct and people who only made a small fix are not on it). You may also add the authors' ORCID.

Archived: https://zenodo.org/record/4565942#.YDkzxy2ZNE4

> * [x]  Please list the DOI of the archived version here.

![image](https://user-images.githubusercontent.com/18452646/109336019-cf807e00-7830-11eb-8fb8-17a10081c591.png)
