### 2021-03-02 #2957 @danielskatz
@ZeyuanSong or @anfederico - I'm suggesting two paper changes in https://github.com/montilab/nf-gwas-pipeline/pull/3, to make the list work (hopefully) and to remove an extra work.  Please merge this and see if it fixes the list (by running `@whedon generate pdf`) or let me know what you disagree with
