### 2021-03-02 #2957 @ZeyuanSong
@danielskatz I forgot to add fundings in the paper, and I just added the fundings in the paper in Github. Could you help update the paper on the journal? Thanks!
