### 2021-03-02 #2957 @ZeyuanSong
@arfon Sorry for the inconvenience caused by my mistake. Could you help edit the paper on the journal?

The updated paper is at [https://github.com/montilab/nf-gwas-pipeline/blob/master/paper/paper.md]. Or you can just substitute the Acknowledgements to "We thank for Harold Bae and Aparna Bhutkar for their help. This work was supported with funding from NIH/NIA: U19AG023122 (Longevity Consortium), and UH2AG064704 (Omics Profiles in Centenarians)."

Many thanks!
