### 2021-03-03 #2957 @arfon
OK, that should be updated now. It may take a few hours for the version on the JOSS site to show as updated due to caching we have in place.
